Moved out logic from PrintModal to Print Helper
